pyproject.toml
[tool.poetry]
name = "bot-tg"
version = "0.1.0"
description = ""
authors = ["Vlasov Dmitry <vlasov.d.e@effective-group.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.7"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins.dotenv]
ignore = "false"
location = ".env.dev"